Program to automatically change playlists based on different executables running.
The purpose of this project is to create playlists of audio files for different executables running. This is mainly used for gaming. For example, if you are playing a fast paced racing game, you might want to listen to playlist A which has upbeat music. If you switch to a slow, stealth-based game, you might want slower, eerie music. This program switches playlists automatically when you switch games.